fix cleanup of config backups, make script more robust
The cleanup of the config backups in /etc/gitlab/config_backup is not
working. The cleanup script uses the wrong path and config backups are
piling up. I fixed the path and replaced ls with find for deleting config
backups. I also changed the file glob to gitlab_config*.tar and
*gitlab_backup.tar files to prevent deletion of other files.

Bug: T274463
Change-Id: I46505706d338491697996e5b1939eab5930d2af8