make prometheus exporters reachable
This change will set all exporters to listen to the public interface
Prometheus was not able to scrape the metrics because they were
listening on localhost only.
Nginx needs additional parameters, otherwise it will return 403 -
Forbidden for /metrics. The integration in the gitlab.rb.j2
template could be better, but I would like to clean that up when
migrating to puppet.

Background:
https://docs.gitlab.com/ee/administration/monitoring/prometheus/#using-an-external-prometheus-server

Bug: T275170
Change-Id: I926a75c0293b3cd7993c1c82d994b6f571ad5c94