gitlab cas: uid instead of CN; add nickname_key
The goal here is that user URLs / @-mentions should equate to shell
name, while the user's display name should correspond to CN.

Without nickname_key, this seems to fall back to cas:user value - the
username the user used to log in to CAS.

Background URLs collected from IRC discussion between jbond and brennen:

  - https://github.com/omniauth/omniauth/wiki/Auth-Hash-Schema
  - https://github.com/tduehr/omniauth-cas3/blob/master/lib/omniauth/strategies/cas3.rb#L39
  - https://apereo.github.io/cas/4.2.x/protocol/CAS-Protocol-Specification.html#257-example-response-with-custom-attributes

Testing:

This has been checked on a fresh GitLab instance by:

  1. Running `sudo gitlab-ctl cleanse` on gitlab-ansible-test
  2. Applying original version of patch without nickname_key
  3. Logging in as "Rando McRandomface"
     - User gets https://gitlab-test.wmcloud.org/RandoMcRandomface
  4. Applying current version of patch with nickname_key
  5. Logging in as Brennen Bearnes
     - User gets https://gitlab-test.wmcloud.org/brennen
  6. Attempting a fresh login with "Rando McRandomface"
     - User is authed as existing RandoMcRandomface user

It _appears_ this could be applied to the production instance with no
loss of data, and future signups would be handled correctly.  We could
then deal with migrating existing users separately.

Bug: T288392
Change-Id: I1cacfd2433637ffac3cd3f3798ea98bc2dc37664